+++ src/gui/pages/initial_page.rs
@@ -8,13 +8,14 @@ use iced::alignment::{Horizontal, Vertical};
 use iced::widget::scrollable::Direction;
 use iced::widget::tooltip::Position;
 use iced::widget::{
-    button, horizontal_space, vertical_space, Button, Column, Container, Row, Rule, Scrollable,
-    Text, TextInput, Tooltip,
+    button, Button, Checkbox, Column, Container, Row, Rule, Scrollable, Space, Text, TextInput,
+    Tooltip,
 };
 use iced::Length::FillPortion;
-use iced::{alignment, Font, Length, Renderer};
+use iced::{alignment, Alignment, Font, Length};
 use pcap::Device;
 
+use crate::gui::components::button::button_open_file;
 use crate::gui::styles::button::ButtonType;
 use crate::gui::styles::container::ContainerType;
 use crate::gui::styles::scrollbar::ScrollbarType;
@@ -22,6 +23,7 @@ use crate::gui::styles::style_constants::{FONT_SIZE_SUBTITLE, FONT_SIZE_TITLE};
 use crate::gui::styles::text::TextType;
 use crate::gui::styles::text_input::TextInputType;
 use crate::gui::styles::types::gradient_type::GradientType;
+use crate::gui::types::export_pcap::ExportPcap;
 use crate::gui::types::message::Message;
 use crate::gui::types::sniffer::Sniffer;
 use crate::networking::types::filters::Filters;
@@ -31,13 +33,16 @@ use crate::translations::translations::{
     address_translation, addresses_translation, choose_adapters_translation,
     ip_version_translation, protocol_translation, select_filters_translation, start_translation,
 };
-use crate::translations::translations_3::port_translation;
-use crate::utils::formatted_strings::get_invalid_filters_string;
+use crate::translations::translations_3::{
+    directory_translation, export_capture_translation, file_name_translation, port_translation,
+};
+use crate::utils::formatted_strings::{get_invalid_filters_string, get_path_termination_string};
+use crate::utils::types::file_info::FileInfo;
 use crate::utils::types::icon::Icon;
 use crate::{ConfigSettings, IpVersion, Language, Protocol, StyleType};
 
 /// Computes the body of gui initial page
-pub fn initial_page(sniffer: &Sniffer) -> Container<Message, Renderer<StyleType>> {
+pub fn initial_page(sniffer: &Sniffer) -> Container<Message, StyleType> {
     let ConfigSettings {
         style,
         language,
@@ -84,6 +89,11 @@ pub fn initial_page(sniffer: &Sniffer) -> Container<Message, Renderer<StyleType>
         )
         .push(Rule::horizontal(40))
         .push(
+            Container::new(get_export_pcap_group(&sniffer.export_pcap, language, font))
+                .height(Length::Fill)
+                .align_y(Vertical::Top),
+        )
+        .push(
             Container::new(button_start(
                 font,
                 language,
@@ -92,14 +102,14 @@ pub fn initial_page(sniffer: &Sniffer) -> Container<Message, Renderer<StyleType>
             ))
             .width(Length::Fill)
             .height(Length::Fill)
-            .align_y(Vertical::Center)
+            .align_y(Vertical::Top)
             .align_x(Horizontal::Center),
         );
 
-    let body = Column::new().push(vertical_space(Length::Fixed(5.0))).push(
+    let body = Column::new().push(Space::with_height(5)).push(
         Row::new()
             .push(col_adapter)
-            .push(horizontal_space(Length::Fixed(30.0)))
+            .push(Space::with_width(30))
             .push(filters_pane),
     );
 
@@ -110,7 +120,7 @@ fn col_ip_buttons(
     active_ip_filters: &HashSet<IpVersion>,
     font: Font,
     language: Language,
-) -> Column<'static, Message, Renderer<StyleType>> {
+) -> Column<'static, Message, StyleType> {
     let mut buttons_row = Row::new().spacing(5).padding([0, 0, 0, 5]);
     for option in IpVersion::ALL {
         let is_active = active_ip_filters.contains(&option);
@@ -122,8 +132,8 @@ fn col_ip_buttons(
                     .vertical_alignment(Vertical::Center)
                     .font(font),
             )
-            .width(Length::Fixed(90.0))
-            .height(Length::Fixed(35.0))
+            .width(90)
+            .height(35)
             .style(if is_active {
                 ButtonType::BorderedRoundSelected
             } else {
@@ -149,7 +159,7 @@ fn col_protocol_buttons(
     active_protocol_filters: &HashSet<Protocol>,
     font: Font,
     language: Language,
-) -> Column<'static, Message, Renderer<StyleType>> {
+) -> Column<'static, Message, StyleType> {
     let mut buttons_row = Row::new().spacing(5).padding([0, 0, 0, 5]);
     for option in Protocol::ALL {
         let is_active = active_protocol_filters.contains(&option);
@@ -157,13 +167,12 @@ fn col_protocol_buttons(
         buttons_row = buttons_row.push(
             Button::new(
                 Text::new(format!("{option} {check_symbol}"))
-                    .width(Length::Fill)
                     .horizontal_alignment(Horizontal::Center)
                     .vertical_alignment(Vertical::Center)
                     .font(font),
             )
-            .width(Length::Fixed(90.0))
-            .height(Length::Fixed(35.0))
+            .width(90)
+            .height(35)
             .style(if is_active {
                 ButtonType::BorderedRoundSelected
             } else {
@@ -189,7 +198,7 @@ fn col_address_input(
     value: &str,
     font: Font,
     language: Language,
-) -> Column<'static, Message, Renderer<StyleType>> {
+) -> Column<'static, Message, StyleType> {
     let is_error = if value.is_empty() {
         false
     } else {
@@ -200,7 +209,7 @@ fn col_address_input(
             .padding([3, 5])
             .on_input(Message::AddressFilter)
             .font(font)
-            .width(Length::Fixed(310.0))
+            .width(310)
             .style(if is_error {
                 TextInputType::Error
             } else {
@@ -224,7 +233,7 @@ fn col_port_input(
     value: &str,
     font: Font,
     language: Language,
-) -> Column<'static, Message, Renderer<StyleType>> {
+) -> Column<'static, Message, StyleType> {
     let is_error = if value.is_empty() {
         false
     } else {
@@ -235,7 +244,7 @@ fn col_port_input(
             .padding([3, 5])
             .on_input(Message::PortFilter)
             .font(font)
-            .width(Length::Fixed(180.0))
+            .width(180)
             .style(if is_error {
                 TextInputType::Error
             } else {
@@ -260,7 +269,7 @@ fn button_start(
     language: Language,
     color_gradient: GradientType,
     filters: &Filters,
-) -> Tooltip<'static, Message, Renderer<StyleType>> {
+) -> Tooltip<'static, Message, StyleType> {
     let mut content = button(
         Icon::Rocket
             .to_text()
@@ -269,8 +278,8 @@ fn button_start(
             .vertical_alignment(alignment::Vertical::Center),
     )
     .padding(10)
-    .height(Length::Fixed(80.0))
-    .width(Length::Fixed(160.0))
+    .height(80)
+    .width(160)
     .style(ButtonType::Gradient(color_gradient));
 
     let mut tooltip = start_translation(language).to_string();
@@ -284,13 +293,12 @@ fn button_start(
         position = Position::FollowCursor;
     }
 
-    Tooltip::new(content, tooltip, position)
+    Tooltip::new(content, Text::new(tooltip).font(font), position)
         .gap(5)
-        .font(font)
         .style(ContainerType::Tooltip)
 }
 
-fn get_col_adapter(sniffer: &Sniffer, font: Font) -> Column<Message, Renderer<StyleType>> {
+fn get_col_adapter(sniffer: &Sniffer, font: Font) -> Column<Message, StyleType> {
     let ConfigSettings { language, .. } = sniffer.configs.lock().unwrap().settings;
 
     let mut dev_str_list = vec![];
@@ -357,4 +365,66 @@ fn get_col_adapter(sniffer: &Sniffer, font: Font) -> Column<Message, Renderer<St
             ))
             .direction(Direction::Vertical(ScrollbarType::properties())),
         )
+}
+
+fn get_export_pcap_group(
+    export_pcap: &ExportPcap,
+    language: Language,
+    font: Font,
+) -> Container<'static, Message, StyleType> {
+    let enabled = export_pcap.enabled();
+    let file_name = export_pcap.file_name();
+    let directory = export_pcap.directory();
+
+    let caption = export_capture_translation(language);
+    let checkbox = Checkbox::new(caption, enabled)
+        .on_toggle(move |_| Message::ToggleExportPcap)
+        .size(18)
+        .font(font);
+
+    let mut ret_val = Column::new().spacing(10).push(checkbox);
+
+    if enabled {
+        let inner_col = Column::new()
+            .spacing(10)
+            .padding([0, 0, 0, 45])
+            .push(
+                Row::new()
+                    .align_items(Alignment::Center)
+                    .spacing(5)
+                    .push(Text::new(format!("{}:", file_name_translation(language))).font(font))
+                    .push(
+                        TextInput::new(ExportPcap::DEFAULT_FILE_NAME, file_name)
+                            .on_input(Message::OutputPcapFile)
+                            .padding([2, 5])
+                            .font(font)
+                            .width(200),
+                    ),
+            )
+            .push(
+                Row::new()
+                    .align_items(Alignment::Center)
+                    .spacing(5)
+                    .push(Text::new(format!("{}:", directory_translation(language))).font(font))
+                    .push(Text::new(get_path_termination_string(directory, 25)).font(font))
+                    .push(button_open_file(
+                        directory.to_owned(),
+                        FileInfo::Directory,
+                        language,
+                        font,
+                        true,
+                        Message::OutputPcapDir,
+                    )),
+            );
+        ret_val = ret_val.push(inner_col);
+        Container::new(ret_val)
+            .padding(10)
+            .width(Length::Fill)
+            .style(ContainerType::BorderedRound)
+    } else {
+        Container::new(ret_val)
+            .padding(10)
+            .width(Length::Fill)
+            .style(ContainerType::BorderedRound)
+    }
 }